requirements.txt
accelerate==0.31.0
aiohttp==3.9.3
aiosignal==1.3.1
alabaster==0.7.13
annotated-types==0.6.0
anyio==4.3.0
arango==0.2.1
arrow==1.3.0
asgiref==3.8.1
attrs==23.2.0
azure-common==1.1.28
azure-core==1.30.1
azure-cosmos==4.7.0
azure-cosmosdb-nspkg==2.0.2
azure-cosmosdb-table==1.0.6
azure-functions==1.19.0
azure-nspkg==3.0.2
babel==2.14.0
backoff==2.2.1
blinker==1.8.2
blis==0.7.11
catalogue==2.0.10
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
confection==0.1.5
cryptography==42.0.7
cymem==2.0.8
dataclasses==0.6
distro==1.9.0
django==3.2.12
djangorestframework==3.13.1
docker==7.1.0
docutils>=0.12,<0.17
filelock==3.13.1
flask==3.0.1
frozenlist==1.4.1
fsspec==2024.2.0
gitdb==4.0.11
gitpython==3.1.42
gql==3.5.0
graphql-core==3.2.3
gunicorn==21.2.0
h11==0.14.0
halo==0.0.31
httpcore==1.0.4
httpx==0.27.0
huggingface-hub==0.23.3
idna==3.6
imagesize==1.4.1
iniconfig==2.0.0
itsdangerous==2.2.0
jinja2==3.1.3
langcodes==3.4.0
language-data==1.2.0
log-symbols==0.0.14
marisa-trie==1.1.1
markdown-it-py==3.0.0
markupsafe==2.1.5
mdurl==0.1.2
mock==5.1.0
mpmath==1.3.0
multidict==6.0.5
murmurhash==1.0.10
networkx==2.8.8
nose==1.3.7
openai==1.12.0
packaging==23.2
pandas==1.4.4
pillow==10.3.0
pluggy==1.4.0
preshed==3.0.9
psutil==5.9.8
pycparser==2.22
pydantic==2.6.2
pydantic-core==2.16.3
pygments==2.17.2
pymupdf==1.24.1
pymupdfb==1.24.1
pytest==8.0.2
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
pyyaml==6.0.1
regex==2023.12.25
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.7.0
safetensors==0.4.2
shellingham==1.5.4
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.0
snowballstemmer==2.2.0
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sphinx>=4.0.0
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp>=1.0.0,<2.0.0

